data_engineering/requirements.txt
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
croniter==0.3.36
dask==2.30.0
decorator==4.4.2
defusedxml==0.6.0
distributed==2.30.1
docker==4.3.1
entrypoints==0.3
HeapDict==1.0.1
idna==2.10
importlib-metadata==3.0.0
importlib-resources==3.3.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-oneofschema==2.1.0
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.0
mypy-extensions==0.4.3
natsort==7.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.3
oauthlib==3.1.0
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pendulum==2.1.2
pickleshare==0.7.5
prefect==0.13.16
prometheus-client==0.9.0
prompt-toolkit==3.0.8
psutil==5.7.3
pycparser==2.20
Pygments==2.7.2
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-box==5.2.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-slugify==4.0.1
pytz==2020.4
pytzdata==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==4.7.7
QtPy==1.9.0
requests==2.25.0
requests-oauthlib==1.3.0
Send2Trash==1.5.0
six==1.15.0
sortedcontainers==2.3.0
SQLAlchemy==1.3.20
tabulate==0.8.7
tblib==1.7.0
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
tweepy==3.9.0
typed-ast==1.4.1
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.5.1
wrapt==1.12.1
zict==2.0.0
zipp==3.4.0
